refactor(core): move checkout redirect as a route
