+++ .changeset/bumpy-waves-behave.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+Refactors redirecting to checkout as a route. This will enable session syncing to happen through a redirect using the sites and routes API.
