+++ core/app/[locale]/(default)/cart/_actions/redirect-to-checkout.ts
@@ -1,28 +1,12 @@
 'use server';
 
-import { BigCommerceGQLError } from '@bigcommerce/catalyst-client';
 import { SubmissionResult } from '@conform-to/react';
 import { parseWithZod } from '@conform-to/zod';
 import { getLocale, getTranslations } from 'next-intl/server';
 import { z } from 'zod';
 
-import { getSessionCustomerAccessToken } from '~/auth';
-import { client } from '~/client';
-import { graphql } from '~/client/graphql';
 import { redirect } from '~/i18n/routing';
 import { getCartId } from '~/lib/cart';
-
-const CheckoutRedirectMutation = graphql(`
-  mutation CheckoutRedirectMutation($cartId: String!) {
-    cart {
-      createCartRedirectUrls(input: { cartEntityId: $cartId }) {
-        redirectUrls {
-          redirectedCheckoutUrl
-        }
-      }
-    }
-  }
-`);
 
 export const redirectToCheckout = async (
   _lastResult: SubmissionResult | null,
@@ -30,8 +14,6 @@ export const redirectToCheckout = async (
 ): Promise<SubmissionResult | null> => {
   const locale = await getLocale();
   const t = await getTranslations('Cart.Errors');
-
-  const customerAccessToken = await getSessionCustomerAccessToken();
 
   const submission = parseWithZod(formData, { schema: z.object({}) });
 
@@ -41,37 +23,5 @@ export const redirectToCheckout = async (
     return submission.reply({ formErrors: [t('cartNotFound')] });
   }
 
-  let url;
-
-  try {
-    const { data } = await client.fetch({
-      document: CheckoutRedirectMutation,
-      variables: { cartId },
-      fetchOptions: { cache: 'no-store' },
-      customerAccessToken,
-    });
-
-    url = data.cart.createCartRedirectUrls.redirectUrls?.redirectedCheckoutUrl;
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error(error);
-
-    if (error instanceof BigCommerceGQLError) {
-      return submission.reply({
-        formErrors: error.errors.map(({ message }) => message),
-      });
-    }
-
-    if (error instanceof Error) {
-      return submission.reply({ formErrors: [error.message] });
-    }
-
-    return submission.reply({ formErrors: [t('failedToRedirectToCheckout')] });
-  }
-
-  if (!url) {
-    return submission.reply({ formErrors: [t('failedToRedirectToCheckout')] });
-  }
-
-  return redirect({ href: url, locale });
+  return redirect({ href: '/checkout', locale });
 };
