+++ core/messages/en.json
@@ -443,7 +443,6 @@
     "Errors": {
       "cartNotFound": "No cartId cookie found.",
       "lineItemNotFound": "Line item not found.",
-      "failedToRedirectToCheckout": "Failed to redirect to checkout.",
       "failedToUpdateQuantity": "Failed to update quantity.",
       "somethingWentWrong": "Something went wrong. Please try again later."
     }
